version 3.6
- new drag and drop using dnd-kit
- fix errors when opening restore from backup window
- larger popup window
- fix some cases where sync would delete all collections
