+++ app/DeleteWithConfirmationButton.js
@@ -20,7 +20,7 @@ const DeleteWithConfirmationButton = (props) => {
         <div className={`slider ${confirmOpen ? 'slider-open' : null}`}>
             <button className="slider-button" disabled={confirmDisabled} data-tip="Delete this group and all its tabs?" data-class="small-tooltip" onClick={handleDelete}>Delete</button>
         </div>
-        <div title={confirmOpen ? 'Cancel' : `Delete group '${props.group.title}'`} className="del slider-del" onClick={toggleSlideConfirm}>
+        <div data-tip={confirmOpen ? 'Cancel' : `Delete group '${props.group.title}'`} className="del slider-del" onClick={toggleSlideConfirm}>
             <MdDeleteForever color="#B64A4A" size="20px" />
         </div>
     </div>
