+++ app/constants.js
@@ -1,0 +1,1 @@
+export const UNDO_TIME = 10;
