+++ chrome/manifest.json
@@ -80,5 +80,5 @@
     ]
   },
   "key": "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAgbFz4tW1ZT+Vmf+jMh5GJOLfHVI2UNoEDaqJKn0ZMC+9+9nePws++SBA9/lkQFjTc9symjYrgkr389ZBsPAtLsGC1D099eVrPeADts3pWYn0KopJjXIMBxcqRmffC7VwpYYCxJ1olACQuE1WHRvjNX4J84xWwCLf7lV1dCjB+viNnRLp2VJ7xKDj6/axdco2x9NVK8/0qRlH1eGe3i956hV+u8DkhF9fgH0sKKQACgJRSUE1fF1Y+FvDZFw7EdYSFTAxSxOCblE+8dw5Kwr0edPd3hWRehTV0bawTW4CYyRVhiICW9UogmDpTCkAgXbhzc8sDjHDF96ChdPPQE8KkwIDAQAB",
-  "version": "3.5.9"
+  "version": "3.6"
 }
