+++ package.json
@@ -53,53 +53,52 @@
     ]
   },
   "dependencies": {
-    "@babel/runtime": "^7.22.15",
-    "@sentry/react": "^7.69.0",
-    "@sentry/tracing": "^7.69.0",
-    "@szhsin/react-menu": "^4.0.3",
+    "@babel/runtime": "^7.26.0",
+    "@dnd-kit/core": "^6.1.0",
+    "@dnd-kit/sortable": "^8.0.0",
+    "@dnd-kit/utilities": "^3.2.2",
+    "@sentry/react": "^8.36.0",
+    "@szhsin/react-menu": "^4.2.2",
     "babel-plugin-transform-runtime": "^6.23.0",
     "javascript-time-ago": "2.3.13",
     "react": "^17.0.2",
     "react-confirm-alert": "2.8.0",
     "react-countdown-circle-timer": "^3.2.1",
     "react-dom": "^17.0.2",
-    "react-icons": "^4.11.0",
-    "react-select": "^5.7.4",
+    "react-icons": "^5.3.0",
+    "react-select": "^5.8.2",
     "react-simple-animate": "^3.5.2",
     "react-simple-snackbar": "^1.1.11",
-    "react-sortablejs": "^6.1.4",
     "react-time-ago": "^7.2.1",
-    "react-tiny-popover": "^7.2.4",
-    "react-tooltip": "^4.5.0",
-    "recoil": "^0.7.7",
-    "sortablejs": "1.15.0"
+    "react-tiny-popover": "^8.1.2",
+    "react-tooltip": "^4.5.1",
+    "recoil": "^0.7.7"
   },
   "devDependencies": {
-    "@babel/core": "^7.22.17",
-    "@babel/plugin-transform-runtime": "^7.22.15",
-    "@babel/preset-env": "^7.22.15",
-    "@babel/preset-react": "^7.22.15",
-    "@sentry/webpack-plugin": "^2.7.1",
+    "@babel/core": "^7.26.0",
+    "@babel/plugin-transform-runtime": "^7.25.9",
+    "@babel/preset-env": "^7.26.0",
+    "@babel/preset-react": "^7.25.9",
+    "@sentry/webpack-plugin": "^2.22.6",
     "@types/jest": "^28.1.8",
-    "@types/sortablejs": "^1.15.2",
     "babel-jest": "^27.5.1",
     "babel-loader": "8.2.5",
     "copy-webpack-plugin": "^10.2.4",
     "css-loader": "^6.8.1",
     "eslint": "^8.49.0",
     "eslint-plugin-jest": "^26.7.0",
-    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react": "^7.37.2",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.5.3",
     "jest": "^27.5.1",
     "jest-environment-jsdom": "^27.5.1",
     "jest-junit": "^14.0.1",
     "jest-transform-stub": "^2.0.0",
-    "jest-webextension-mock": "^3.8.9",
+    "jest-webextension-mock": "^3.9.0",
     "react-test-renderer": "^17.0.2",
     "style-loader": "^3.3.3",
-    "webextension-polyfill": "^0.10.0",
-    "webpack": "^5.88.2",
+    "webextension-polyfill": "^0.12.0",
+    "webpack": "^5.95.0",
     "webpack-cli": "^5.1.4",
     "webpack-extension-manifest-plugin": "^0.8.0"
   }
