+++ static/index.css
@@ -53,7 +53,7 @@ body {
   padding:0;
   margin:0;
   border:0;
-  width: 550px;
+  width: 650px;
   min-height: 520px;
   overflow-x:auto;
   background-color:var(--bg-color);
