+++ static/index.js
@@ -1,7 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import * as Sentry from "@sentry/react";
-import { Integrations } from "@sentry/tracing";
 import App from "../app/App";
 import "./index.css";
 import { RecoilRoot } from 'recoil';
@@ -10,7 +9,7 @@ import { browser } from '../static/globals';
 
 Sentry.init({
     dsn: "https://315246baa1ed462883bab2ecd7507290@o975143.ingest.sentry.io/5931101",
-    integrations: [new Integrations.BrowserTracing()],
+    integrations: [Sentry.browserTracingIntegration()],
     ignoreErrors: [
         'ResizeObserver loop limit exceeded',
     ],
